Automatically resume connection
git-svn-id: svn://localhost/trunk@121 1a4c2bc4-540c-48e5-8c71-7a1e3271b4ae

+++ source/textoperations/TextEditor.cpp
@@ -32,7 +32,7 @@
 #define FONTSIZE    18
 
 void *LoadFile(char *filepath, int size);
-unsigned int GetFileSize(const char *filename);
+off_t GetFileSize(const char *filename);
 
 /**
  * Constructor for the TextEditor class.
